test: delete hook binding file and move it to custom hook test
